+++ R/raster.R
@@ -106,30 +106,25 @@ setAs("stars_proxy", "Raster", function(from) {
 	raster::brick(unlist(from))
 })
 
-get_terra_levels = function(x, min_v) {
+get_terra_levels = function(x) {
 # create factor levels, as used by stars, from SpatRaster levels in a data.frame
 # see https://github.com/r-spatial/stars/pull/484
-	IDs = x[[1]]
-	if (any(IDs < 0))
+	x = x[order(x[[1]]), ] # sort table on level
+	levels = x[[1]]
+	if (any(levels < 0))
 		stop("negative IDs in SpatRaster levels not supported")
-	categories = x[[2]]
-	if (min(IDs) == 0) {
-		if (min_v == 0) # shift all:
-			IDs = IDs + 1
-		else { # remove ID 0:
-			categories = categories[IDs != 0]
-			IDs = IDs[IDs != 0]
-		}
-	}
-	ct = rep("_", max(IDs))
-	ct[IDs] = categories
-	make.unique(ct)
+	ex = setdiff(0:max(levels), levels)
+	exclude = rep(FALSE, max(levels) + 1)
+	exclude[ex + 1] = TRUE # 0-based vector
+	list(levels = levels, labels = x[[2]], exclude = exclude)
 }
 
 #' @name st_as_stars
 #' @param ignore_file logical; if \code{TRUE}, ignore the SpatRaster object file name
+#' @param as_attributes logical; if \code{TRUE} and \code{.x} has more than one layer, load these as separate attributes rather than as a band or time dimension (only implemented for the case where \code{ignore_file} is \code{TRUE})
 #' @export
-st_as_stars.SpatRaster = function(.x, ..., ignore_file = FALSE) {
+st_as_stars.SpatRaster = function(.x, ..., ignore_file = FALSE,
+			  as_attributes = all(terra::is.factor(.x))) {
 	if (!requireNamespace("terra", quietly = TRUE))
 		stop("package terra required, please install it first") # nocov
 
@@ -139,13 +134,13 @@ st_as_stars.SpatRaster = function(.x, ..., ignore_file = FALSE) {
 	src = terra::sources(.x, bands=TRUE)
 
 	attr_name = basename(src$source[1])
-	if (attr_name == "")
-		attr_name = "values"
 
 	if (!ignore_file && all(src$source != "")) {
 	# there can be multiple files, but only the first one is used here.
 	# perhaps a warning should be given; better would be to iterate over "sid"
 	# but you might have a situation where some sources are filenames and others are not
+		if (attr_name == "")
+			attr_name = "values"
 		lst = vector("list", length(unique(src$sid)))
 		for (i in unique(src$sid)) {
 			file = unique(src$source[src$sid == i])
@@ -178,33 +173,39 @@ st_as_stars.SpatRaster = function(.x, ..., ignore_file = FALSE) {
 
 		setNames(ret, attr_name)
 	} else { # ignore_file TRUE:
+		if (terra::nlyr(.x) > 1 && as_attributes) {
+			ret = do.call(c, lapply(seq_len(terra::nlyr(.x)), function(i) st_as_stars(.x[[i]], ignore_file = TRUE)))
+			if (!is.null(names(.x)))
+				names(ret) = names(.x)
+			return(ret) # RETURNS
+		}
+		if (attr_name == "") {
+			if (all(names(.x) == ""))
+				attr_name = "values"
+			else
+				attr_name = paste(names(.x)[1], collapse = ".")
+		}
 		v = terra::values(.x, mat = FALSE)
-		dim(v) = dim(.x)[c(2,1,3)]
+		dimv = dim(v) = dim(.x)[c(2,1,3)]
 		if (all(terra::is.factor(.x))) {
 			if (length(terra::levels(.x)) > 1)
 				warning("ignoring categories/levels for all but first layer")
-			if ((min_v <- min(v, na.rm = TRUE)) == 0) # +warn here?
-				v = v + 1
-			l = terra::levels(.x)[[1]]
-			if (inherits(l, "data.frame"))
-				l = get_terra_levels(l, min_v)
+			if (inherits(l <- terra::levels(.x)[[1]], "data.frame"))
+				l = get_terra_levels(l)
+			else
+				stop("terra levels should return a list of data.frame's; pls update terra")
 			colors = try(rgb(terra::coltab(.x)[[1]], maxColorValue = 255), silent = TRUE)
 			if (inherits(colors, "try-error") || length(colors) == 0)
 				colors = NULL
-			else if (length(colors) == length(levels) + 1) # remove last color?
-				colors = colors[-length(colors)]
-			v = structure(v, class = "factor", levels = as.character(l), colors = colors)
+			v = structure(factor(as.vector(v), levels = l$levels, labels = l$labels),
+					dim = dimv, colors = colors, exclude = l$exclude)
 		}
 		dimensions = list(
-			x = create_dimension(from = 1, to = dim(v)[1], offset = e[1],
+				x = create_dimension(from = 1, to = dim(v)[1], offset = e[1],
 							 	delta = (e[2]-e[1])/dim(v)[1], refsys = st_crs(terra::crs(.x))),
-			y = create_dimension(from = 1, to = dim(v)[2], offset = e[4],
+				y = create_dimension(from = 1, to = dim(v)[2], offset = e[4],
 							 	delta = (e[3]-e[4])/dim(v)[2], refsys = st_crs(terra::crs(.x))))
 		dimensions$band = create_dimension(values = names(.x))
-	#	l = if (length(names) > 1)
-	#		setNames(list(v), deparse(substitute(.x), 50))
-	#	else
-	#		setNames(list(v), names(.x)[1])
 		ret = st_as_stars(list(v), dimensions = create_dimensions(dimensions, get_raster()))
 		if (dim(ret)[3] == 1)
 			ret = adrop(ret, 3)
@@ -300,11 +301,33 @@ st_as_raster = function(x, class, ...) {
 	dxy = attr(d, "raster")$dimensions
 	stopifnot(all(dxy %in% names(d)))
 	bb = st_bbox(x)
-	values = if (is.factor(x[[1]])) {
-		structure(x[[1]], dim = NULL)
-	} else {
-		as.vector(x[[1]]) # would convert factor into character
-	}
+	levels = NULL
+	coltab = vector("list", length(x))
+	values = if (all(sapply(x, is.factor))) {
+			ex = attr(x[[1]], "exclude")
+			if (is.null(ex) || class != "SpatRaster")
+				structure(merge(x)[[1]], dim = NULL) # return the factor
+			else {
+				v = vector("list", length(x))
+				levels = vector("list", length(x))
+				for (i in seq_along(v)) {
+					ex = attr(x[[i]], "exclude")
+					ix = which(!ex) - 1 # 0-based index
+					n = as.numeric(structure(x[[i]], dim = NULL)) # factor -> numeric
+					v[[i]] = ix[n]
+					levels[[i]] = data.frame(IDs = ix, categories = levels(x[[i]]))
+					if (!is.null(ct <- attr(x[[i]], "colors"))) {
+						coltab[[i]] = t(col2rgb(rep("#000000", length(ex)), alpha = TRUE))
+						coltab[[i]][which(!ex),] = t(col2rgb(ct, alpha = TRUE))
+					}
+				}
+				do.call(c, v)
+			}
+		} else  {
+			if (any(sapply(x, is.factor)))
+				warning("mix of factor and non-factor attributes: all factor levels are ignored")
+			as.vector(merge(x)[[1]])
+		}
 	if (class == "SpatRaster"){
 		third = setdiff(names(d), dxy)
 		b = terra::rast(nrows = dim(x)[ dxy[2] ], ncols=dim(x)[ dxy[1] ],
@@ -312,8 +335,11 @@ st_as_raster = function(x, class, ...) {
 						nlyrs = ifelse(length(dim(x)) == 2, 1, dim(x)[third]),
 						crs = x_crs$wkt)
 		terra::values(b) = values
-		if (all(vapply(x, is.factor, FUN.VALUE = logical(1)))) {
-			terra::coltab(b) = lapply(x, function(x) t(col2rgb(attr(x, "colors"), alpha = TRUE)))
+		if (!is.null(levels)) {
+			levels(b) = levels
+			if (!all(sapply(coltab, is.null)))
+				for (i in seq_len(terra::nlyr(b)))
+					terra::coltab(b, layer = i) = coltab[[i]]
 		}
 		if (length(dim(x)) != 2){
 			z = seq(d[[third]])