+++ R/rasterize.R
@@ -142,6 +142,8 @@ guess_raster = function(x, ...) {
 #' if (require(plm, quietly = TRUE)) {
 #'   data(Produc, package = "plm")
 #'   st_as_stars(Produc, y_decreasing = FALSE)
+#'  data(Produc, package = "plm")
+#'  st_as_stars(Produc, y_decreasing = FALSE)
 #' }
 st_as_stars.data.frame = function(.x, ..., dims = coords, xy = dims[1:2], y_decreasing = TRUE, coords = 1:2) {