+++ tests/aggregate.R
@@ -8,8 +8,9 @@ d = st_dimensions(station = st_set_crs(st_as_sfc(stations), 4326), time = dates)
 blocks = st_make_grid(st_as_sfc("POLYGON ((5.871619 47.26986, 15.03811 47.26986, 15.03811 55.05653, 5.871619 55.05653, 5.871619 47.26986))", crs = 4326),
 	n = c(3,3))
 
-(aq = st_as_stars(list(PM10 = air), dimensions = d))
-(a = aggregate(aq, blocks, mean, na.rm = TRUE))
+print(aq <- st_as_stars(list(PM10 = air), dimensions = d))
+print(a <- aggregate(aq, blocks, mean, na.rm = TRUE))
+}
 
 # adapted from ?read_stars:
 m = array(1:720, dim = c(x = 10, y = 12, t = 6)) # named dim