+++ tests/testthat/test_st_as_stars.R
@@ -3,6 +3,7 @@ context("st_as_stars tests")
 options(rgdal_show_exportToProj4_warnings = "none")
 
 test_that("basic st_as_stars", {
+  skip_if_not_installed("ncdfgeom")
   skip_on_os("solaris")
   skip_if_not_installed("ncdfgeom")
 
@@ -57,7 +58,6 @@ test_that("basic st_as_stars", {
 
 test_that("st_as_stars.Raster", {
   skip_if_not_installed("raster")
-
   tif = system.file("tif/L7_ETMs.tif", package = "stars")
   x = raster::raster(tif)
   stars_obj <- st_as_stars(x)